% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{a_paper}
\alias{a_paper}
\title{Get details of a single paper}
\usage{
a_paper(
  id = "649def34f8be52c8b66281af98ae884c09aef38b",
  args = "citations/?fields=year,fieldsOfStudy,citationCount"
)
}
\arguments{
\item{id}{semanticscholar id, doi, pubmed,...}

\item{get}{what to get - hardcode now}
}
\description{
Get details of a single paper
}
