% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_paper}
\alias{get_paper}
\title{Get papers from search_author results}
\usage{
get_paper(search_author, id)
}
\arguments{
\item{search_author}{\code{\link[paper]{search_author}} output}

\item{id}{authorId identify from \code{\link[paper]{search_author}} output.}
}
\description{
Get papers from search_author results
}
