% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_as_tibble}
\alias{list_as_tibble}
\title{List to tibble handling NULL element}
\usage{
list_as_tibble(x)
}
\arguments{
\item{x}{a list with possible NULL element}
}
\value{
a tibble with NULL in elements replaced by NA
}
\description{
List to tibble handling NULL element
}
