% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring-distribution-plot.R
\name{offspring_distribution_plot}
\alias{offspring_distribution_plot}
\title{Plot offspring distribution}
\usage{
offspring_distribution_plot(outbreak.dataset, title = NULL,
  include.unobserved.as.zero = FALSE)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{title}{Plot title. Default = NULL}

\item{include.unobserved.as.zero}{Boolean detailing whether idividuals who have not
yet ended their infection by the end of the practical are included and as R0 = 0 occurences.
Default = FALSE.}
}
\description{
\code{offspring_distribution_plot} takes output of \code{outbreak_dataset_read} and
plots offspring distribution. Returns ggplot object for plotting.
}

