\name{papeR-package}
\alias{papeR-package}
\alias{papeR}
\docType{package}
\title{
  A Toolbox for Writing Pretty Papers and Reports
}
\description{
  A toolbox for writing knitr, Sweave or other LaTeX- or markdown-based
  reports and to prettify the output of various estimated models.
}
\details{
  \tabular{ll}{
    Package: \tab papeR\cr
    Type: \tab Package\cr
    Version: \tab 1.0-0\cr
    Date: \tab 2015-12-01\cr
    License: \tab GPL-2\cr
  }

  Version 1.0-0 is based on a completely refactored code base. Some
  functions from previsous versions are deprecated. New functions to
  create summary tables exist (see \code{\link{summarize}}. The package now
  also provides a vignette and was extensively tested using
  \code{testthat}.

  For news and changes see \code{news(package = "papeR")}.
}
\author{
  Benjamin Hofner

  Maintainer: Benjamin Hofner <Benjamin.Hofner@fau.de>
}

\keyword{package}
