% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper}
\alias{paper}
\title{Paper}
\usage{
paper(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{some data}

\item{width}{width of component}

\item{height}{height of component}

\item{elementId}{component element ID}

\item{opacity}{opacity of image}

\item{source}{path to image}
}
\description{
Some description
}
