% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramater-boxplots-plot.R
\name{paramater_boxplots_plot}
\alias{paramater_boxplots_plot}
\title{Plot 4x1 boxplots of epidemiological parameters}
\usage{
paramater_boxplots_plot(outbreak.dataset, title)
}
\arguments{
\item{outbreak.dataset}{Outbreak dataset outputted by \code{outbreak_dataset_read}}

\item{title}{Plot title}
}
\description{
\code{paramater_boxplots_plot} takes output of \code{outbreak_dataset_read} and
plots boxplot of epidemiological parameter observations
}

