% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{plot.an_author}
\alias{plot.an_author}
\title{Plot method for an_author class}
\usage{
\method{plot}{an_author}(an_author, give_data = FALSE, txt_seed = 1)
}
\arguments{
\item{an_author}{an \code{\link[paper]{an_author}} object}

\item{give_data}{give me data instead of the plot}

\item{txt_seed}{seed to label text use in \code{ggrepel}}
}
\description{
Produce a similar plot to PlumX metric, but the numbers of papers cited the
reference paper by fields.
}
\details{
\itemize{
\item Size is the number of papers times the median citation counts of the
respective paper
\item Older median year of the papers give are more transparent.
}
}
