% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_author}
\alias{search_author}
\title{Search an author by name}
\usage{
search_author(
  query = "adam smith ",
  args = "name,affiliations,aliases,url,papers.title,papers.year,papers.url",
  limit = 10
)
}
\arguments{
\item{query}{author name as character string, space allow}

\item{limit}{number of result to return if there are more than this limit}
}
\description{
Search an author by name
}
