% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{update_api}
\alias{update_api}
\title{Update API url if needed}
\usage{
update_api(new_url)
}
\arguments{
\item{new_url}{new API url, eg "https://api.semanticscholar.org/graph/v1/"}
}
\description{
In case API changes
}
