% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{view_web}
\alias{view_web}
\title{Open author website (if exists) in default browser}
\usage{
view_web(search_author, id)
}
\arguments{
\item{search_author}{\code{\link[paper]{search_author}} output}

\item{id}{authorId identify from \code{\link[paper]{search_author}} output.}
}
\description{
Open author website (if exists) in default browser
}
