% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{an_author}
\alias{an_author}
\title{Get details of a single paper}
\usage{
an_author(
  id = "1741101",
 
    args = "paperCount,citationCount,hIndex,papers.year,papers.url,papers.citationCount,papers.fieldsOfStudy",
  level = c("me_only", "me_and_cited_me")
)
}
\arguments{
\item{id}{semanticscholar id, doi, pubmed,...}

\item{level}{get the author paper's details only or also details of their
cited papers?}

\item{get}{what to get - hardcode now TODO: deal with whatever field requested}
}
\description{
Get details of a single paper
}
