\name{papeR-package}
\alias{papeR-package}
\alias{papeR}
\docType{package}
\title{
  A toolbox for writing Sweave or other LaTeX-based papers and reports
}
\description{
  A toolbox for writing Sweave or other LaTeX-based papers and reports
}
\details{
  \tabular{ll}{
    Package: \tab papeR\cr
    Type: \tab Package\cr
    Version: \tab 0.3\cr
    Date: \tab 2013-08-07\cr
    License: \tab GPL-2\cr
  }
}
\author{
  Benjamin Hofner

  Maintainer: Benjamin Hofner <Benjamin.Hofner@fau.de>
}

\keyword{package}
